+++ src/packages/map/ui/MapAnchor.test.tsx
@@ -1,6 +1,7 @@
 // SPDX-FileCopyrightText: 2023 Open Pioneer project (https://github.com/open-pioneer)
 // SPDX-License-Identifier: Apache-2.0
-import { MapContainer, MapPadding } from "./MapContainer";
+import { MapContainer } from "./MapContainer";
+import { MapPadding } from "../api";
 import { createServiceOptions, setupMap, waitForMapMount } from "@open-pioneer/map-test-utils";
 import { PackageContextProvider } from "@open-pioneer/test-utils/react";
 import { render, waitFor } from "@testing-library/react";