+++ src/packages/selection/CHANGELOG.md
@@ -1,1 +1,20 @@
 # @open-pioneer/selection
+
+## 0.1.0
+
+### Minor Changes
+
+-   ee7c2d4: Update runtime version.
+-   61d3e0e: Initial release.
+-   0456500: Add interface `BaseFeature` to Map API.
+-   35be8ef: Introduction of `SelectionSourceFactory` for creating selection sources of an OpenLayers VectorLayer.
+
+### Patch Changes
+
+-   Updated dependencies [ee7c2d4]
+-   Updated dependencies [a582e5e]
+-   Updated dependencies [0456500]
+-   Updated dependencies [762e7b9]
+    -   @open-pioneer/map@0.3.0
+    -   @open-pioneer/notifier@0.3.0
+    -   @open-pioneer/react-utils@0.2.1